refactor(perf): split /inner-circle + /ai-world into server shell + motion island
Wave 28. 17th + 18th in the pattern series. Server shells delegate
to client Shell components for motion content. Inner-circle keeps
its layout.tsx-based FAQ + Organization + Service + Breadcrumb
JSON-LD; ai-world layout.tsx supplies metadata.
